// src/Pages/Authenticated/Profile/Types/IProfile.tsx
import IAccountInfo from "Utils/Types/IAccountInfo";
import IBasicInfo from "Utils/Types/IBasicInfo";
import IContactInfo from "Utils/Types/IContactInfo";
import IPersonalInfo from "Utils/Types/IPersonalInfo";

export default interface IProfile {
    accountInfo: IAccountInfo;
    basicInfo: IBasicInfo;
    contactInfo: IContactInfo;
    personalInfo: IPersonalInfo;
}